+++ packages/web-components/fast-foundation/src/virtual-list/index.ts
@@ -1,3 +1,4 @@
+export * from "./virtual-list.options.js";
 export * from "./virtual-list.js";
 export * from "./virtual-list.template.js";
 export * from "./virtual-list-item.js";
